+++ src/sdk/PnP.Core.Admin/Model/Teams/Internal/TeamCreator.cs
@@ -22,6 +22,7 @@ namespace PnP.Core.Admin.Model.Teams
             stopwatch.Start();
 
             bool isProvisioningComplete = false;
+            bool teamAlreadyExisted = false;
             var retryAttempt = 1;
             Exception lastException = null;
             do
@@ -44,17 +45,27 @@ namespace PnP.Core.Admin.Model.Teams
                     // Log and eat exception here
                     context.Logger.LogWarning(PnPCoreAdminResources.Log_Warning_ExceptionWhileCreatingTeamForGroup, teamToCreate.GroupId, ex.ToString());
                     lastException = ex;
+
+                    // Leave the retry loop if the Team already existed
+                    if ((ex is MicrosoftGraphServiceException graphServiceException) && graphServiceException.Error is MicrosoftGraphError microsoftGraphError)
+                    {
+                        // Conflict: means the Team already exists
+                        if (microsoftGraphError.HttpResponseCode == 409)
+                        {
+                            teamAlreadyExisted = true;
+                        }
+                    }                    
                 }
 
                 retryAttempt++;
             }
-            while (!isProvisioningComplete && retryAttempt <= creationOptions.MaxStatusChecks);
+            while (!(isProvisioningComplete || teamAlreadyExisted) && retryAttempt <= creationOptions.MaxStatusChecks);
 
             stopwatch.Stop();
 
             context.Logger.LogDebug($"Elapsed: {stopwatch.Elapsed:mm\\:ss\\.fff} | Finished");
 
-            if (isProvisioningComplete)
+            if (isProvisioningComplete || teamAlreadyExisted)
             {
                 return await GetPnPContextWithTeamAsync(context, teamToCreate.GroupId).ConfigureAwait(false);
             }