+++ src/sdk/CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Admin library: switched to the new multi-geo aware site enumaration endpoint [jansenbe - Bert Jansen]
 - Fixed `LoadListDataAsStream` paging issue when item count was a multiple of the page size #1296 [jansenbe - Bert Jansen]
 - Made it easier to set a `IField` internal name [jansenbe - Bert Jansen]
+- Leave 'Teamify' loop when Team already existed #1319 [jansenbe - Bert Jansen]
 
 ## [1.10.0]